migration token structure
