+++ src/lib.rs
@@ -16,7 +16,7 @@ pub use codegen::{
 };
 pub use dom::DomNode;
 pub use error::{MarkdownError, Result};
-pub use lexer::{Lexer, Position, Token};
+pub use lexer::{Lexer, Position, token::Token};
 pub use parser::ParserConfig;
 pub use performance::{ParallelConfig, PerformanceOptimizer, ZeroCopyStr};
